+++ src/test/java/racingcar/input/utils/racingtrytime/RacingTryTimeValidatorUtilsTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import racingcar.input.utils.racingtrytime.exception.NotNaturalNumberRacingTimeException;
+import racingcar.input.utils.racingtrytime.validators.RacingTryTimeValidatorUtils;
 
 public class RacingTryTimeValidatorUtilsTest {
